-double to float replacement -minor misc tweaks
svn path=/trunk/FreeOrion/; revision=6516

+++ universe/ResourceCenter.h
@@ -56,7 +56,7 @@ public:
 protected:
     void            Init();                                                         ///< initialization that needs to be called by derived class after derived class is constructed
 
-    double          ResourceCenterNextTurnMeterValue(MeterType meter_type) const;   ///< returns estimate of the next turn's current values of meters relevant to this ResourceCenter
+    float           ResourceCenterNextTurnMeterValue(MeterType meter_type) const;   ///< returns estimate of the next turn's current values of meters relevant to this ResourceCenter
     void            ResourceCenterResetTargetMaxUnpairedMeters();
     void            ResourceCenterClampMeters();
